+++ src/CookieTailor.tsx
@@ -1,12 +1,14 @@
 import Cookies from "js-cookie";
+import { v4 as uuidv4 } from 'uuid';
 import React, { Component, CSSProperties } from "react";
 import FooterTailor from "./components/FooterTailor";
 import { ConditionalWrapper } from "./components/ConditionalWrapper";
 import { CookieTailorProps, defaultTailorProps } from "./CookieTailor.props";
 import { CookieTailorState, defaultState } from "./CookieTailor.state";
 import { SAME_SITE_OPTIONS, VISIBILITY_OPTIONS } from "./types";
-import { getCookieTailorValue, getLegacyCookieName } from "./utilities";
+import { getTailorCookieValue, getLegacyCookieName } from "./utilities";
 import "./css/out/rct_style.css";
+import { defaultCookiePrefix } from "./constants";
 
 export class CookieTailor extends Component<CookieTailorProps, CookieTailorState> {
   public static defaultProps = defaultTailorProps;
@@ -116,7 +118,7 @@ export class CookieTailor extends Component<CookieTailorProps, CookieTailorState
    */
   getCookieValue() {
     const { cookieName } = this.props;
-    return getCookieTailorValue(cookieName);
+    return getTailorCookieValue(cookieName);
   }
 
   /**
@@ -147,6 +149,22 @@ export class CookieTailor extends Component<CookieTailorProps, CookieTailorState
       window.removeEventListener("scroll", this.handleScroll);
     }
   };
+
+  getDefaultCookieId = () => {
+    const cookieName = `${defaultCookiePrefix}id`;
+    return getTailorCookieValue(cookieName);
+  }
+
+  createDefaultCookie = () => {
+    if(!this.getDefaultCookieId()) {
+      const randomId = uuidv4();
+      const cookieName = `${defaultCookiePrefix}id`;
+      const cookieCreation = `${defaultCookiePrefix}created`;
+      const createdDateAsString = new Date().toISOString();
+      this.setCookie(cookieName, randomId);
+      this.setCookie(cookieCreation, createdDateAsString);
+    }
+  }
 
   render() {
     // If the bar shouldn't be visible don't render anything.
@@ -197,6 +215,7 @@ export class CookieTailor extends Component<CookieTailorProps, CookieTailorState
     }
 
     myStyle.backgroundColor = tailorColors.background;
+    this.createDefaultCookie();
     return (
       <ConditionalWrapper
         condition={overlay}