+++ src/types/tailorLabels.ts
@@ -42,6 +42,8 @@ interface MainLabels {
   buttonAllow: string;
   buttonDefault: string;
   buttonClose: string;
+  defaultId: string;
+  defaultCreation: string;
   descriptionPrivacyLink: string;
   descriptionPrivacyTitle: string;
   descriptionStart: string;